refactor: trim code health follow-up debt
