+++ src/routes/+layout.ts
@@ -1,5 +1,4 @@
-import { initAppServices, getAppServices } from '$lib/app/composition';
-import { runStartupMigrations } from '$lib/app/startup-migrations';
+import { initAppServices } from '$lib/app/composition';
 import { injectAnalytics } from '@vercel/analytics/sveltekit';
 import { browser } from '$app/environment';
 
@@ -11,6 +10,5 @@ injectAnalytics({ mode: 'auto' });
 export async function load() {
 	if (browser) {
 		await initAppServices();
-		runStartupMigrations(getAppServices());
 	}
 }
